% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hipercow_utils.R
\name{check_package_in_file}
\alias{check_package_in_file}
\title{Check if a Package Exists in pkgdepends.txt}
\usage{
check_package_in_file(pkg, file)
}
\arguments{
\item{pkg}{A character string specifying the package name to check.}

\item{file}{A character string specifying the path to the \code{pkgdepends.txt} file.}
}
\value{
A logical value: \code{TRUE} if the package is found in the file, \code{FALSE} otherwise.
}
\description{
This helper function checks whether a given package is listed in the \code{pkgdepends.txt} file.
}
\keyword{internal}
