% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet_param_gen.R
\name{combine_itn_data}
\alias{combine_itn_data}
\title{Combine ITN datasets}
\usage{
combine_itn_data(file_paths)
}
\arguments{
\item{file_paths}{A named list of file paths to the CSV files.}
}
\value{
A combined data frame of all ITN data.
}
\description{
Combine ITN datasets
}
