% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderly_utils.R
\name{execute_checks}
\alias{execute_checks}
\title{Execute Parameter Set Checks}
\usage{
execute_checks(start, end, parallel)
}
\arguments{
\item{start}{The starting index of the parameter sets to check.}

\item{end}{The ending index of the parameter sets to check.}

\item{parallel}{Logical. If \code{TRUE}, the checks are run in parallel using multiple cores; otherwise, they are run sequentially.}
}
\value{
A list of results for each parameter set, including success flags and any errors.
}
\description{
This function executes the \code{check_parameter_set} function for a range of parameter sets, either sequentially or in parallel.
}
\keyword{internal}
