% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hipercow_utils.R
\name{extract_packages}
\alias{extract_packages}
\title{Extract Package Names from R Script}
\usage{
extract_packages(file)
}
\arguments{
\item{file}{A character string specifying the path to an R file.}
}
\value{
A character vector of package names used in the script.
}
\description{
This helper function extracts package names from \code{library()} and \code{require()} calls in an R script.
}
\keyword{internal}
