% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hipercow_utils.R
\name{find_missing_packages}
\alias{find_missing_packages}
\title{Find Missing Packages in R Scripts}
\usage{
find_missing_packages(src_dir, pkg_file)
}
\arguments{
\item{src_dir}{A character string specifying the path to the directory containing R scripts.}

\item{pkg_file}{A character string specifying the path to the \code{pkgdepends.txt} file.}
}
\value{
A character vector of missing package names.
}
\description{
This helper function processes all R files in a specified directory,
identifies packages that are not listed in the \code{pkgdepends.txt} file,
and returns a list of those missing packages.
}
\keyword{internal}
