% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{flatten_list}
\alias{flatten_list}
\title{Flatten Nested Lists (Internal)}
\usage{
flatten_list(x, parent_key = "")
}
\arguments{
\item{x}{List to flatten}

\item{parent_key}{Parent key for naming}
}
\value{
Flattened list
}
\description{
Flatten Nested Lists (Internal)
}
