% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hipercow_utils.R
\name{hipercow_check_pkgdepends}
\alias{hipercow_check_pkgdepends}
\title{Check for Missing Packages in pkgdepends.txt}
\usage{
hipercow_check_pkgdepends()
}
\value{
NULL. The function may prompt the user to take actions such as adding missing packages
to \code{pkgdepends.txt} or reprovisioning the hipercow directory.
}
\description{
This function checks all R files in the "src" directory for \code{library()} and \code{require()} calls,
identifies any packages that are not listed in the \code{pkgdepends.txt} file, and prompts the user to
add any missing packages.
}
