% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hipercow_utils.R
\name{list_r_files}
\alias{list_r_files}
\title{List All R Files Recursively}
\usage{
list_r_files(dir)
}
\arguments{
\item{dir}{A character string specifying the directory to search for R files.}
}
\value{
A character vector containing the full paths of all R files found.
}
\description{
This helper function lists all R files within a specified directory, including subdirectories.
}
\keyword{internal}
