% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet_param_gen.R
\name{load_itn_data}
\alias{load_itn_data}
\title{Load and transform ITN data}
\usage{
load_itn_data(file_path, net_type)
}
\arguments{
\item{file_path}{The path to the CSV file.}

\item{net_type}{The type of net to label the data.}
}
\value{
A data frame with the loaded and transformed data.
}
\description{
Load and transform ITN data
}
