% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderly_utils.R
\name{orderly_prod}
\alias{orderly_prod}
\title{Run Parameter Set Checks}
\usage{
orderly_prod(
  indices = 1:10000,
  verbose = TRUE,
  parallel = TRUE,
  store_output = FALSE
)
}
\arguments{
\item{indices}{A vector of indices of the parameter sets to check. Default is \code{1:10000}.}

\item{verbose}{Logical. If \code{TRUE}, detailed output of the process is provided (default is \code{TRUE}).}

\item{parallel}{Logical. If \code{TRUE}, the checks are run in parallel using multiple cores; otherwise, they are run sequentially (default is \code{TRUE}).}

\item{store_output}{Logical. If \code{TRUE}, the function will store the IDs of the successful checks and the parameter sets that encountered errors (default is \code{FALSE}).}
}
\value{
A list containing the success summary, error summary, and optionally the IDs and error parameter sets if \code{store_output} is \code{TRUE}.
}
\description{
This function runs checks on a vector of parameter sets and provides a detailed report on the results.
It can execute the checks either sequentially or in parallel.
}
