% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderly_utils.R
\name{prod}
\alias{prod}
\title{Run Parameter Set Checks}
\usage{
prod(start = 1, end = 10000, verbose = TRUE, parallel = TRUE)
}
\arguments{
\item{start}{The starting index of the parameter sets to check (default is 1).}

\item{end}{The ending index of the parameter sets to check (default is 10,000).}

\item{verbose}{Logical. If \code{TRUE}, detailed output of the process is provided (default is \code{TRUE}).}

\item{parallel}{Logical. If \code{TRUE}, the checks are run in parallel using multiple cores; otherwise, they are run sequentially (default is \code{TRUE}).}
}
\value{
None. This function prints a summary of the results and detailed information if \code{verbose} is \code{TRUE}.
}
\description{
This function runs checks on a range of parameter sets and provides a detailed report on the results.
It can execute the checks either sequentially or in parallel.
}
