% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet_param_gen.R
\name{save_itn_data}
\alias{save_itn_data}
\title{Save ITN parameters to an RDS file}
\usage{
save_itn_data(data, output_path)
}
\arguments{
\item{data}{The data frame to save.}

\item{output_path}{The path to the output RDS file.}
}
\description{
Save ITN parameters to an RDS file
}
